Enhance Normal.ps1 to prevent deletion of non-empty directories; update Archive.ps1 to ensure proper execution of Updater with cleaner and installer parameters.

+++ Updaters/Apps/Cleaners/Normal.ps1
@@ -61,6 +61,10 @@ function HandleDirectory {
         $RemoveItem.Invoke($_);
     }
 
+    if (Get-ChildItem -LiteralPath $DirectoryInfo.FullName -ErrorAction SilentlyContinue) {
+        return;
+    }
+
     Remove-Item -LiteralPath $DirectoryInfo.FullName -Force
 }
